% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\docType{data}
\name{motley_palette}
\alias{motley_palette}
\title{Motley default qualitative palette (16 colours)}
\format{
An object of class \code{character} of length 16.
}
\usage{
motley_palette
}
\description{
Motley default qualitative palette (16 colours)
}
\keyword{datasets}
